EH-747: Moved AiemminHankittuYhteinenTutkinnonOsa.ts to a new folder
